requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
dnspython==2.6.1
docutils==0.21.2
ecdsa==0.19.0
email_validator==2.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fiona==1.9.6
fonttools==4.51.0
geopandas==0.14.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
icecream==2.1.3
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jose==1.0.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mypy==1.10.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openpyxl==3.1.2
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pandoc==2.3
parso==0.8.4
passlib==1.7.4
pillow==10.3.0
plumbum==1.8.3
ply==3.11
polars==0.20.25
prettyprint==0.1.5
proglog==0.1.10
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
setuptools==70.0.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
